pyproject.toml
[project]
name = "pydoommap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "pillow>=11.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[tool.pytest.ini_options]
pythonpath = [
    "src"
]
